+++ modelopt/torch/utils/__init__.py
@@ -23,6 +23,7 @@ from .import_utils import *
 from .list import *
 from .logging import *
 from .loss_mask import *
+from .mlflow import *
 from .network import *
 from .perf import *
 from .regex import *